🐛 Improve command fuzzy matching
This fixes the panic caused when using `did you mean` with a command
longer than the arguments. This also improves the fuzzy matching by
looking for the closest matching set of arguments in the ones provided,
and using the leftovers as arguments to the selected command.

Fixes https://github.com/XaF/omni/issues/11