✨ Add support for `version_file` for golang up step
The `version_file` parameter can now be used in the map of the
golang up step; when used, this value should aim at the `go.mod`
file and omni will try and read the golang version from it.

If no version is found in the file, or if no file is found, omni
will raise an error when reaching that step.

Closes https://github.com/XaF/omni/issues/17