✨ Add self updater for omni (#26)
As was done in the ruby version, omni will offer to auto-update
when a new version is published. This by default will be queried
to the user, but can be configured (`path_repo_updates.self_update`)
to never happen (`false`), happen automatically (`true`) or keep
asking (`ask`, which is the default).

The update will work for binary installs, having the binary download
the new version tarball, check against the sha256 for integrity, and
replace itself. It will also work for brew installs, where the binary
will simply call `brew upgrade xaf/omni/omni` to trigger an upgrade
of itself.

Closes https://github.com/XaF/omni/issues/24