🐛 Fix duplicate `tool:version` after `omni up`
When the same `tool:version` was used multiple times in the same
`omni up`, we didn't check that it was the same, which led to
loading the dynamic environment for it multiple times. This fixes
that by checking if the `tool:version` is already in the dynamic
environment before adding it.

Fixes https://github.com/XaF/omni/issues/23