✨ Add `worktree` configuration parameter
This allows to not mandate using the `OMNI_GIT` environment
variable. If the environment variable is set, though, it will
override the configuration of `worktree`.

The default value for `worktree` corresponds to the default
value that was used for `OMNI_GIT`, i.e. ~/git or $GOPATH/src
if any of those exists, or revert to ~/git by default.