✨ self updater can now update the user configuration
Instead of just asking yes/no, it will now suggest extra options
to always auto-update or never do. This also changes the behavior
of the `false` value for `self_update`, where it will still show
to the user that there is an update, even if not doing any update.
The new value `nocheck` has been added to entirely skip checking
for updates.