✨ Add support to use omni in regular directories
Until now, omni could only handle dependencies in a git repository
with a configured remote, as it would allow having a repository id
to identify the work directory.

This adds a new `workdir` structure which tries to use the git
repository remote if available, but can fallback to a generated
id for the repository stored in `.omni/id`.

This generated id is not intended for share, as it would be pretty
unsafe, allowing someone to specify an id in a repository that people
trust and then reusing it in an untrusted, malicious repository. To
avoid this, we use the unique machine id to append a hash to the
generated id. Upon finding an invalid `.omni/id`, its value will
simply be ignored.