✨ Add support for `omni cd --locate`
This allows to return on stdout the location of a repository.
This is particularly useful when trying to write scripts and commands
depending on another repository, since omni will be able to locate it
and return the path in a safe way, as long as users of the command
also use omni.