✨ Split the cache into multiple files
This allows for time-sensitive caches, such as the up_environments cache
for omni's dynamic environment, to be separated from more
user-experience and resource handling caches such as the asdf_operation
one. This means that time-sensitive features can only read a smaller
file when necessary to access the cache, vs. loading and parsing the
full content of other cache keys that are unimportant to the behavior.

Fixes https://github.com/XaF/omni/issues/131