feat: ✨ Add suggest_config and suggest_clone fingerprints
This allows to take fingerprints of the contents of `suggest_config` and
`suggest_clone` when bootstrapping a repository. No matter what the user
decides to do, it saves that fingerprint and compares it on future runs,
allowing to track any change to those from the repository point of view,
which would warrant suggesting the changes again to the use.

This also introduces a new `up_command/auto_bootstrap` boolean
configuration parameter to now default to `omni up --bootstrap` when
running `omni up`, and thus automatically suggesting changes whenever
the repository suggestions get updated.

Closes https://github.com/XaF/omni/issues/66