feat: ✨ Add support for 'packages'
This might include **breaking changes** in terms of the behaviors of
some commands.

Packages are a stable, non-development copy of a git repository
that can be used for providing commands in a stable way. Those
repositories will be cloned in a different (data) directory than the
development repositories, and will be loaded in the omnipath as a simple
package reference, e.g.

```
path:
  append:
    - package: git@github.com:XaF/omni
      path: cmd
```

This will be handled the same way as the regular omnipath is, except
that it won't be in your development directories. Use this approach for
repositories that provide you commands, but that you don't expect to
develop with. You can also use this as to have a stable copy of a
repository, even if that repository also exists in your development
directories.

The following configuration is affected by this change:
- `suggest_clone` entries could take `handle` and `args`, and
  can now also take `clone_type` as one of `package` or `worktree`,
  affecting how the package is going to be suggested to the user
  when suggesting the clone; this defaults to `package` if
  not specified

The following commands are also evolved from this change:
- `omni up` now defaults to clone suggested repositories as packages,
  with a new option `p` when being asked about cloning or not those
  repositories. The `y` option is still available to clone in the
  worktree. When splitting the list with `s`, a second question will be
  asked the user, allowing to individually indicate how to clone each
  selected repository.
- `omni clone` gets a new `--package` option to clone a repository as a
  package. It will default to clone, as usual, in the development
  worktree corresponding to the matching organization.
- `omni cd` gets a new `--[no-]include-packages` option to indicate
  whether to include or not packages when calling `omni cd`. This
  defaults to not including packages when calling `omni cd` directly as
  it is expected to be used for development purposes, but will default
  to including packages when calling `omni cd` with the `--locate`
  option as it is expected to be used in programmatic uses. The option
  allows to force one behavior or the other.
- Repository location, using `omni scope` and `omni cd` will now
  consider the omnipath in priority to locating the repository in order
  of organizations.
- `omni tidy` will identify and clone packages in the omnipath that are
  not available locally

Closes https://github.com/XaF/omni/issues/16