feat: ✨ `omni help` now folds subcommands
`omni help` can become pretty long as commands and categories get
added. This changes the help message to fold subcommands into their
parents when more than one command is provided for a specific path.
This also adds support for `omni help <command>` where `<command>`
does not exist but is a path leading to existing command(s). In such
situation, this will show the subcommands valid for that path.
This also allows for `omni <command>` where `<command>` does not exist
but is a path leading to existing command(s) to show a `Did you mean`
message suggesting the given subcommands.

Closes https://github.com/XaF/omni/issues/20