feat: ✨ add new `config path switch` command (#199)
This command allows to easily switch a repository between worktree and
package modes. The command usage is as follows:

    omni config path switch [--package|--worktree] [repo]

If none of `--package` or `--worktree` is provided, the command will
simply toggle the source of the repository. If `repo` is not provided,
the current directory will be used.

The command errors out if:
- No `repo` is provided and the current directory is not a repository
- The `repo` is not found, or cannot be resolved
- The repository is not currently cloned (at all)

Closes https://github.com/XaF/omni/issues/196