feat: ✨ Add `config bootstrap` command
This will walk the caller through setting up the configuration of
the `worktree`, the `repo_path_format` and the first organizations
if any repository is available.

Omni will also check if a configuration file is available and, if
none, will trigger the `config bootstrap` command for getting the
user started with omni.

Closes https://github.com/XaF/omni/issues/178