feat: ✨ add shell integration during bootstrap (#221)
Omni's installation require setting up the shell integration. As part of
https://github.com/XaF/omni/issues/179 we made an install script
available to make it easy to install omni in a few instants, and as part
of https://github.com/XaF/omni/pull/209 we added the `config bootstrap`
command to bootstrap omni's configuration. This now adds automatic
addition of omni's shell integration to the rc file of the caller, which
will make it even easier to get started with omni.

Closes https://github.com/XaF/omni/issues/210