feat: ✨ save log files when subprocesses fail (#243)
For some operations such as `custom` operations when calling `omni up`,
the previous message was a bit cryptic. It will now print simply the
exit code and a path to a saved log file. Most subprocesses that are
called in the background for a better UX will hence have their logs
saved in a way that makes it easier to understand what happened when an
execution failed.