feat(updates): ✨ add support for background updates
This will make it smoother to use omni as only the required updates
will be done synchronously (i.e. if the command being run is from
the path to be updated), while other updates will be done
asynchronously in the background.

For asynchronous updates, the log file will be saved in case of any
error, and printed the next time the `env` hook is being called.

Closes https://github.com/XaF/omni/issues/170