fea(updater): ✨ pre-auth to hosts to avoid multiple auth queries
When doing updates, multiple `git pull` are being done in parallel,
which can be bothersome if using a password on the ssh key or a ssh
agent with authentication, since with multiple parallel queries,
the user will have to give multiple approvals.

This adds a pre-auth step when doing updates that will run a
`ls-remote` on each individual host found in the repositories to
trigger the authentication process. Failed auth will lead to
skipping updates for repositories in that host.