fix(path commands): make the exec path accurate
With the handling of aliases for symlinks, the execution path was
lost and the first encountered path was the one being executed for
any of the aliases to one command. This prevented being able to
use `$0` in bash scripts for instance, or any equivalent in any
language when writing an omni command.

This fixes that problem by making sure that aliases that have their
own execution path, which is the case for symlinks, are being executed
from that path, making `$0` (and equivalents) usable.

Fixes https://github.com/XaF/omni/issues/161