feat(path commands): ✨ Add support for metadata files (#253)
The idiomatic way to add omni metadata to a command is to use metadata
headers. However, those are not available for binary files which, so
far, required using a shell wrapper that would allow using headers.

This introduces metadata files, which can be colocated with the
executable files to provide the metadata to those files. Those can be
used for binary files, but also for readable files if preferred (the
idiomatic way staying metadata headers for those, however).

Closes https://github.com/XaF/omni/issues/27