feat(homebrew operation): ✨ update PATH with installed formulae
This allows to make sure that any tool installed for a given
repository is made available in the context of that repository.
Prior to this change, it was expected that the homebrew bin path
was loaded in the PATH environment variable by the user.

This slightly changes the update cache format to store the formulae
bin paths. A new `--no-cache` parameter has been added to `omni up`
in order to allow running `omni up` without taking the cache into
account. This option only works for homebrew for now.

Closes https://github.com/XaF/omni/issues/256