fix(homebrew operation)!: 🐛 require explicit `cask` mention
With `brew` automatically falling back to cask when a formula
is not found, this prevented omni to handle operations for casks
in a different way than for formulae, for things that only work
for formulae (e.g. `brew --prefix`). This changes that by passing
an explicit `--formula` or `--cask` when performing actions.

BREAKING CHANGE: casks that did not have the `cask` explicit mention
in `up:` operations will stop working. Simple fix is to prefix them by
`cask:` if the cask name is provided inline (e.g. `- 1password-cli`
should become `- cask: 1password-cli`.