feat(up tools): ✨ better cleanup for data paths
https://github.com/XaF/omni/pull/277 introduces venv creation and
was handling the deletion as part of the venv setup, which did not
allow to handle multiple versions of python at the same time,
except while using `auto`.

This fixes that by moving the data path cleanup to the cleanup
process that handles removing unused versions at the end of
an `omni up` or `omni down`.

This also means that it naturally handles the removal of the work
directory data path as part of the `omni down` process.

Closes https://github.com/XaF/omni/issues/281