feat(python operation): ✨ Add support for pip install
This supports specifying one or more requirements files to be
installed with `pip install -r` along with the python installation.

If the python installation is for multiple, isolated directories,
each venv will get the `pip install -r` calls for each of the files.

If `auto` is passed to the `pip:` parameter, the default
`requirements.txt` file will be looked for in each matching directory.

If the python version is `auto` and no `pip` parameter is specified,
the `requirements.txt` file will be looked for in each matching
directory.

Closes https://github.com/XaF/omni/issues/279