feat(golang operation): ✨ improve golang isolation to maintain shared cache
In the previous commit, the isolation was not taking into account
the go mod cache location, which is by default going to
gopath[0]/pkg/mod (with gopath defaulting to ~/go if unset).
This means that the cache was being duplicated as many times as
repositories.

This takes care of that by either:
- maintaining the GOMODCACHE env var as already set, or
- setting the GOMODCACHE env var to gopath[0]/pkg/mod if
  gopath is set or ~/go/pkg/mod otherwise

This makes sure that the go mod cache stays shared while keeping
isolation between working directories, allowing `go install` to have a
different target per isolation, which means that incompatible versions
of the same binary can be installed by separate working directories.