feat(packages): ✨ use a separate cache of dependencies
Packages can coexist with the worktree directory, and this could
lead to development on a directory to override the package
dependencies, or the other way around.

Packages should thus have a separate set of dependencies, making
sure that dependencies are kept both for the package and worktree
versions of a work directory, allowing them to coexist.

This handles that by changing the default id of work directories
to be prefixed by `package#` for packages. The trust check is still
handled through the simple, package-agnostic work directory id.

Closes https://github.com/XaF/omni/issues/278