fix(dynenv): 🐛 unset properly list environment variables
List environment variables were set back to the empty string instead of
being completely unset. This was an issue as empty strings-set
environment variables might not have the same behavior as unset
environment variables.

The issue manifested for the `GOPATH` variable since the recent changes
to handle `GOPATH` as a list in the dynenv.

This fixes that by having a new list operation 'create', which is used
when a list is being appended or prepended to for an environment
variable that is currently unset.