refactor(configuration): ♻️ improvements to the dump and duration handling (#337)
`omni status --config` now dumps the configuration as it is being loaded
by omni, instead of dumping it before full interpretation. This also
changes how the default configuration is loaded, avoiding to duplicate
the default values of each parameter: default configuration can fully be
loaded from an empty object.

Durations in the configuration now accept human-readable formats such as
`1d`, `1w`, `1h`, etc. to make it more accessible.

Finally, this adds configuration parameters for the cache of `asdf` and
`homebrew` operations, allowing the users to specify how long to cache
those information for.