build: 👷 try and make builds a bit more parallel (#338)
Tests for a specific build do not require builds of other versions to
have finished. This creates a reusable workflow that takes care of
building a specific version for a specific suffix, and running the
relevant tests for that version if required.

Given that tests are generally run on faster builds (x86_64), this
should accelerate the whole CI process as those will be able to be ran
in parallel of other processes finishing building binaries.