feat(autocompletion): ✨ consider parent commands
Autocompletion in special cases where we had a subcommand was
not completing for the parent command unless it was clear that
we are not in the path of the subcommand.

For instance, adding an `omni cd blah` command was leading
`omni cd <tab>` to always autocomplete by `blah`, since it
would find the subcommand.

This makes it so, as long as we are not sure we are in the
subcommand path, the parent command gets a chance to
autocomplete too. This is done by calling the autocompletion
on the parent command if found that we are in the process of
autocompleting "one element after" the parent command.

Closes https://github.com/XaF/omni/issues/342