feat(dynenv): ✨ show message when 'omni up' should be run
This allows users to know when the current work directory
has an updated `up` configuration since `omni up` was last
run, or has it available but `omni up` was never run yet.

The notifications will come as the prompt will print, the
same way the dynamic environment update notifications appear,
and use the configuration file(s) modification times but
also a hash of the parameters that are pertinent for the
up configuration, such that changing any other parameters
in the configuration will not trigger a notification. The
only parameters considered at this time are `up`,
`suggest_config` and `suggest_clone`.

This also comes with two configuration options to be able
to disable any of those messages if desired:
- `notify_workdir_config_updated` to enable/disable
  notifications when `omni up` has already been run but
  we detected an update to the `up` configuration
- `notify_workdir_config_available` to enable/disable
  notifications when an `up` configuration is available
  but `omni up` has never been run.

Closes https://github.com/XaF/omni/issues/346