feat: add `prompts` and templating for config and clone suggestions (#418)
Omni helps having an easy onboarding process by using an initial
repository and handling clone and configuration suggestions. One element
that was lacking was the possibility to adjust which configuration
parameters to suggest setting or which repositories to suggest cloning.

This adds support for a new configuration parameter `prompts` which
allows prompting the user and caching the answers. The prompts answer
can be scoped to the repository or to the organization in order to avoid
asking over and over the same questions.

This also adds support for templating of the `suggest_config` and
`suggest_clone` options, including using values obtained from the
prompts.

Closes https://github.com/XaF/omni/issues/162