feat(up): ✨ add dependencies handling and 'or' and 'and' operations
Some of the tooling installed through `asdf` requires dependencies
to already be installed on the system. Those dependencies are stable
and known in advance, as long as one of the known package managers
that omni supports is available on the system.

This adds support so that when installing tools backed by `asdf`,
the right dependencies will be setup on the system using either
`brew` or `nix`. This should be extended in the future as support
for `apt` and `dnf` is added back.

This also introduces new `or` and `and` operations for the `up`
configuration, where `or` allows to indicate that one of the listed
operations must succeed (`or` is marked unavailable and gracefully
skipped if all underneath operations are unavailable), and will
handle operations in the order they are defined, stopping at the
first successful one. The `and` operation requires all of the
listed operations to succeed, and is particularly useful when
combined to an `or` operation.