fix(up hash): 🐛 detect when `env` parameter has changed
We now support a hash of the `up` configuration to indicate to the
user when a new call to `omni up` should be performed. This now also
takes into account the use of the `env` parameter, as it is also being
setup in the dynamic environment and thus requires a refresh when
updated.

Fixes https://github.com/XaF/omni/issues/397