fix(up env cache): 🐛 add default deserialization
The up environment cache deserialization for environment variables
had a bug since we didn't specify the default value for the operation
type if not specified, while at the same time skipping serialization
if the value was the default `set` value.

This fixes that by adding the `default` marker so serde knows to
deserialize into the default value if none specified.