feat(sudo): ✨ handle password requests during `omni up` (#456)
Password requests during `omni up` were pretty cryptic, as hanging the
process but not giving more details about what happened.

This fixes that by taking advantage of the `SUDO_ASKPASS` feature, as
well as `GIT_ASKPASS` and `SSH_ASKPASS` for that matter. This makes it
work automatically with `brew` commands but also with anything that
could contain sudo calls as long as they are performed by using `sudo
-A`.

By default, if omni has been called in interactive mode, this will ask
for the user password on the command line where the omni up command is
running. If omni has been called in non-interactive mode (e.g.
background updates), this will try to use `osascript` on MacOS.
Equivalents for Linux should be added in a follow-up (e.g. `pinentry`
?).

Closes https://github.com/XaF/omni/issues/372
Closes https://github.com/XaF/omni/issues/160