feat(up): ✨ add the `any` operation and `preferred_tools` config
This operation works like `or` but handles user preference where
`or` only considers operation order. If any tools are defined in
`preferred_tools`, these would be considered in priority before
defaulting to the rest of the available operations.

This allows for a better handling of dependencies by using the
tools that the user favor, even if other tooling is available
in the system.

Closes https://github.com/XaF/omni/issues/464