fix(spinners): 🐛 new lines are now ensured properly after a spinner
Previous approach was not keeping track of the length of lines,
which in turn led to newlines being printed even though it was not
required. Now we keep track of the length of the last line printed
in a spinner, and only add a line return if it went over the
expected line length.