feat(askpass): ✨ add configuration for the askpass handling
This allows for users to decide if they prefer to skip entirely
askpass handling or if they wish to use the GUI when possible
even if the password could be asked in the command line directly.

The GUI support is for now only available on MacOS, as depending
on `osascript` to show the dialog.