feat(shims): ✨ add a shims directory with all dynamic binaries (#490)
Some tools cannot work directly with the dynamic environment, such as
some IDEs. Adding a shims directory that dynamically gets updated on
`omni up` allows for a better handling of those IDEs, as they can just
fetch the binary from the PATH.

The shims script will simply load omni's environment for the work
directory they are being called from, and call the corresponding binary
in the PATH. If that PATH does not provide that binary, it will lead to
an error. Otherwise, it will return the first matching version of the
corresponding binary.

This also adds `omni config reshim` command to update all shims if
necessary. This shouldn't be necessary most of the time though, as shims
are being regenerated automatically on `omni up`.

Closes https://github.com/XaF/omni/issues/492