feat(github release): ✨ add more compact configuration handling
GitHub release operation required to add an entry for each release,
which made it verbose both to write the configuration and during
`omni up`. This fixes that by allowing to specify multiple versions,
either as a list or as a map of repository => repository config.

When used with multiple versions, this will show as a single entry
that will summarize the installed releases at the end.