fix(self update): 🐛 force-refresh tap before upgrading omni
If people are disabling homebrew auto-updates, when running
`brew upgrade xaf/omni/omni` there will be no new version to
upgrade to, leaving omni in its previous version.

This fixes that by calling `git pull` in the repository of
omni's tap first, so that we ensure the tap is up-to-date
before calling upgrade.

This also disables other auto-updates of homebrew during the
omni self-update.

Closes https://github.com/XaF/omni/issues/495