fix(self update): 🐛 check version after successful update
The homebrew update was failing when homebrew auto update is disabled
but there was no indicator this was the case, with omni reporting that
the update went through.

This fixes that by checking the version of omni after an upgrade by
calling `omni --version` and checking that the version installed
corresponds to the version we were trying to install.