feat(github release): ✨ try refreshing releases cache if none match (#504)
Previously, the cache was kept for the duration even if no release
matched the requirements. With that change, if the cache was used and no
release matches the requirements, the cache will be refreshed and the
releases evaluated a second time.

This allows to satisfy bound requirements (e.g. specific version number)
even if the change to the requirement is recent and the cache has been
refreshed recently for the user.

Closes https://github.com/XaF/omni/issues/486