feat(asdf): ✨ add suppport for arbitrary asdf tools
A large number of `asdf` tools are available that omni can take
advantage of right away, but this was not available as requiring
specific tool definition to work.
This changes that by changing the default behavior when a tool
is unknown, and considering it's an asdf plugin by default.

This also adds support for the `url` parameter to any asdf tool,
which will override the repository used for that tool. If this
does not correspond to the default tool URL, it will keep the
tool installation separate from one that wouldn't provide an
`url` or would provide a different one.

Closes https://github.com/XaF/omni/issues/494