feat(homebrew): ✨ auto-extract taps from formulas
Homebrew supports auto-tapping formulas using the long format
`tap-owner/tap-name/formula`, which will automatically tap the
formula when calling `brew install`. However, this won't allow
for omni to keep track of what's being tapped and what needs to
be cleaned up.

To solve that problem, omni will now automatically extract the
tap when using that format, so that taps can be automatically
untapped when not used anymore (if omni triggered the initial
tapping).

Closes https://github.com/XaF/omni/issues/516