feat(cleanup): ✨ add delay before removing unused resources
Resources cleanup is done as soon as a resource is not in use.
However, when switching between branches, this might lead to a
lot of resources being cleaned-up and reinstalled, etc.

This adds a grace period between a resource not being used
anymore and it being cleaned up; it defaults to 1 week, but
can be changed to `0` to restore previous behavior, or to
any other discretionary period of time.

Individual configuration options are available depending on the
resource type, so that they can each behave differently:
- `cache`/`asdf`/`cleanup_after` for the cleanup of asdf versions
- `cache`/`homebrew`/`cleanup_after` for the cleanup of homebrew
  taps and formulae
- `cache`/`github_release`/`cleanup_after` for the cleanup of
  GitHub releases

Closes https://github.com/XaF/omni/issues/435