feat(update): ✨ ask before updating on command not found (#522)
`omni` was automatically triggering a synchroneous update on command not
found (if due) in order to make sure to run the command if available.

The new `path_repo_updates`/`on_command_not_found` option in the
configuration can be set to either one of `true`, `false` or `ask`. If
set to `ask`, which is the default, it will ask the user before running
any sync update -- Note that answering 'no' will still trigger updates,
but in the background. If set to `true`, sync updates will be triggered
without any question. If set to `false`, no update will be triggered at
all.

Closes https://github.com/XaF/omni/issues/482