feat(github release): ✨ fallback to installed version if download error
If there is any error during download of a release, we now fallback to
an installed version matching the criteria. This is similar to what was
done for `asdf` operations, but applied to GitHub releases.

For instance, if I am installing version `2`, and the latest version
fitting with `2` is `2.0.9` and I fail downloading it, if `2.0.8` is
already installed, since it fits with my requirement of `2`, ommi
will automatically fallback to that version.

This allows to avoid losing a working dev environment in case the
internet connectivity is having issues, for instance.

Closes https://github.com/XaF/omni/issues/526