feat(github release): ✨ allow for custom asset name matching
This adds an `asset_name` parameter that can take a list of
globs (one per line), for either positive or negative (starting
with `!`) matching. The first matching pattern is then applied
to indicate if an asset is matching or not (if positive, asset
is matching, if negative, asset is not matching, even if a latter
pattern would say the opposite).

This also adds the `skip_os_matching` and `skip_arch_matching`
booleans in the configuration to avoid matching on platform and
architecture respectively. If those parameters are set, any asset
would be downloaded (as long as matching `asset_name` if specified).

Closes https://github.com/XaF/omni/issues/479