fix: 🐛 `env` config parameter of type `path` now resolves properly
The `env` parameters of type `path` were keeping the configuration
file in the resolves relative path, which was obviously not working
and there were no directory of the name of the config file.

This is fixed by resolving the parent of the config file path, which
is thus the directory in which the config file is, and applying the
relative path resolution from there.

Fixes https://github.com/XaF/omni/issues/564