feat(repo_path_format): ✨ can be configured per organization
`repo_path_format` is used to define the format of the path to
repositories inside a worktree. Until now, this was only available
as a global configuration option, while it was possible to define
a different worktree per organization.

This fixes that by adding a `repo_path_format` in the per-org
configuration, which will apply to the specific organization only.

Closes https://github.com/XaF/omni/issues/561