feat: Add `--local` global flag
This new flag makes the current omni process lookup for commands
in the current directory first. It also follows any `suggest_config`
that exists in the repository configuration, even if those suggestions
have not been applied to the user configuration.

Note that this flag does not get inherited when `omni` is then called
from a command called through `omni --local`. The absence of propagation
is done on purpose so that `omni --local` is _required_ for prioritizing
local commands over global ones. We do however set a `OMNI_LOCAL_LOOKUP`
environment variable that can be consumed by subcommands if needed, but
that is ignored by `omni` itself.

Fixes https://github.com/XaF/omni/issues/577