feat(github release): support checksum validation
When running `omni up`, GitHub Releases will now try and look for
checksums assets to verify the downloads. It is possible to
configure the behavior to disable the verification, require it,
verify against a static string, or specify patterns to look for
the asset(s) containing the checksums for a given download.

Closes https://github.com/XaF/omni/issues/539