feat(nodejs): ✨ handle engine versions and package installs (#617)
When using `package.json`, Node can define a number of `engines` and the
relative versions that are needed for it. This was not directly
supported while we were already looking at `package.json` to figure out
the version of node to be installed.

This now adds support to ensure that `npm`, `pnpm`, `yarn`, `electron`
and other engines are installed in the required versions for the current
environment. This also creates an isolation path for node to run `npm
install -g` and install tools only for the current work directory.

This also adds support for running `npm install`, `pnpm install` or
`yarn install` automatically depending on the lock files that are
present in the directory, as well as the `engines` listed in the
`package.json` file.

Both those behaviors default to be enabled, but can be disabled by
setting `install_engines` and `install_packages` to `false` when setting
up the `node` operation.

Closes https://github.com/XaF/omni/issues/293
Closes https://github.com/XaF/omni/issues/294