feat(path commands): ✨ allow to parse python argparse's --help
Python ArgParse returns data is a well known format which can thus
be parsed to extract the information from a command instead of
requiring updated omni headers. This leads to a slightly slower
load time when trying to show a given command's help, but will
in turn simplify keeping commands help up to date.

This is capable of extracting:
- The command description
- The command usage
- The different arguments (positional, option, etc.) and their
  description

It will also override the command's `-h` and `--help` when this
mode is enabled to default to show omni's help, since the command
itself won't be able to request omni to takeover the help by itself,
as it needs its help to be available for omni to consume it.