feat(up): synchronize 'omni up' operations
Until now, `omni up` could be run multiple times in parallel, this
also meant that if a repository is already being updated in the
background, it would be updated a second time synchronously by calling
`omni up`. This changes that by allowing any new `omni up` to attach
to a currently-running `omni up` operation.

This effectively allows to move all updates to be background-only
updates, as it is now possible for an up operation to run in the
background, but to be attached to the current process if necessary.

This still needs to handle:
- up operations with different parameters
- serializing up and down operations (i.e. down should _require_ the
  lock, while `up` can attach if the lock is already taken)
- validate that the last `up`-ed commit is the right one after an
  attached `up` operation that required the repository to be updated,
  in case the attachment was do to an up operation that didn't require
  the repository to be updated