feat(github release): ✨ configurable github auth token
The auth token was previously only pulled from the `gh` tool, by
calls to `gh auth token` which weren't configurable.

This changes that by allowing to configure things at different levels.
First, it can be configured in the user and/or system configuration by
using the new `github:` configuration section:

```yaml
github:
  auth:
    token: <token value>
```

This can also be a list, with matching for the repository and/or the
used API URL, e.g.:

```yaml
github:
  auth:
    - repo:
        glob: XaF/*
      skip: true  # Skip authentication
   - gh  # Use `gh` by default
```

Note that if going through the list, no match is found, it will default
to use the `gh` command line. To skip all authentication by default,
you can use `- skip` as the last value of the list for instance (without
`repo` or `api_url` filter.

Second, it is possible to configure on a per-github-release basis, in
the `up` configuration for a repository:

```yaml
up:
  - github-releases:
      XaF/omni:
        version: latest
        auth:
          token: xxx  # To use a token
      omnicli/xxx:
        auth: skip
```

This allows to apply this setting for that entry, which will take
precedence over the global configuration.

Closes https://github.com/XaF/omni/issues/670
Closes https://github.com/XaF/omni/issues/512